# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Projects Service API
  description: |
    Multi-tenant project and task management API for BengoBox platform.

    ## Authentication
    All endpoints (except health checks) require a valid JWT token in the Authorization header.

    ## Multi-tenancy
    All resource endpoints are scoped to a tenant via the `{tenantID}` path parameter.
  version: 1.0.0
  contact:
    name: BengoBox Platform Team
    email: platform@bengobox.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.bengobox.com/api/v1
    description: Production
...
paths:
...
  /{tenantID}/projects:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      summary: List projects
      description: Returns a paginated list of projects for the tenant
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - name: owner_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create project
      description: Creates a new project in the tenant
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /{tenantID}/projects/{projectID}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/ProjectID'
    get:
      summary: Get project
      description: Returns a single project by ID
      tags: [Projects]
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update project
      description: Updates an existing project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete project
      description: Deletes a project and all associated tasks
      tags: [Projects]
      responses:
        '204':
          description: Project deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from auth-service
  parameters:
    TenantID:
      name: tenantID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant identifier
    ProjectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Project identifier
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip
  schemas:
    ProjectStatus:
      type: string
      enum: [active, planning, on_hold, completed, cancelled]
      description: Project lifecycle status
...
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Website Redesign
        description:
          type: string
          maxLength: 2000
        status:
          $ref: '#/components/schemas/ProjectStatus'
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        budget:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        metadata:
          type: object
          additionalProperties: true
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        status:
          $ref: '#/components/schemas/ProjectStatus'
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        budget:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        metadata:
          type: object
          additionalProperties: true
    ProjectResponse:
      type: object
      required: [id, tenant_id, name, status, currency, owner_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        budget:
          type: number
          format: double
        currency:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    ListProjectsResponse:
      type: object
      required: [projects, total, limit, offset]
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ValidationErrorResponse:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
security:
  - BearerAuth: []